Create Dig.java
Write a program that accepts an integer and checks whether it consists of three digits. If it does, the program should check if all the digits are unique. An appropriate output should be produced at the end. Please do not use any external libraries or data structures that were not discussed in our classes.